+++ newrelic/hooks/component_djangorestframework.py
@@ -1,8 +1,11 @@
+import sys
+
 from newrelic.common.object_wrapper import (wrap_function_wrapper,
         function_wrapper)
 from newrelic.api.transaction import current_transaction
 from newrelic.api.function_trace import FunctionTrace
 from newrelic.common.object_names import callable_name
+
 
 def _nr_wrapper_APIView_dispatch_(wrapped, instance, args, kwargs):
     transaction = current_transaction()
@@ -22,7 +25,8 @@ def _nr_wrapper_APIView_dispatch_(wrapped, instance, args, kwargs):
     else:
         handler = view.http_method_not_allowed
 
-    view_func_callable_name = getattr(view, '_nr_view_func_callable_name', None)
+    view_func_callable_name = getattr(view, '_nr_view_func_callable_name',
+            None)
     if view_func_callable_name:
         if handler == view.http_method_not_allowed:
             name = '%s.%s' % (view_func_callable_name,
@@ -34,8 +38,21 @@ def _nr_wrapper_APIView_dispatch_(wrapped, instance, args, kwargs):
 
     transaction.set_transaction_name(name)
 
+    # catch exceptions handled by view.handle_exception
+    view.handle_exception = _nr_wrapper_APIView_handle_exception_(
+            view.handle_exception, request)
+
     with FunctionTrace(transaction, name):
         return wrapped(*args, **kwargs)
+
+
+def _nr_wrapper_APIView_handle_exception_(handler, request):
+    @function_wrapper
+    def _handle_exception_wrapper(wrapped, instance, args, kwargs):
+        request._nr_exc_info = sys.exc_info()
+        return wrapped(*args, **kwargs)
+    return _handle_exception_wrapper(handler)
+
 
 @function_wrapper
 def _nr_wrapper_api_view_decorator_(wrapped, instance, args, kwargs):
@@ -49,14 +66,17 @@ def _nr_wrapper_api_view_decorator_(wrapped, instance, args, kwargs):
 
     return view
 
+
 def _nr_wrapper_api_view_(wrapped, instance, args, kwargs):
     decorator = wrapped(*args, **kwargs)
     decorator = _nr_wrapper_api_view_decorator_(decorator)
     return decorator
 
+
 def instrument_rest_framework_views(module):
     wrap_function_wrapper(module, 'APIView.dispatch',
             _nr_wrapper_APIView_dispatch_)
+
 
 def instrument_rest_framework_decorators(module):
     wrap_function_wrapper(module, 'api_view',