+++ newrelic/hooks/framework_aiohttp.py
@@ -69,14 +69,14 @@ class NRTransactionCoroutineWrapper(ObjectProxy):
 
             self._nr_transaction = txn
 
-            if txn._settings:
+            if txn.enabled:
                 txn.__enter__()
                 txn.drop_transaction()
 
         txn = self._nr_transaction
 
         # transaction may not be active
-        if not txn._settings:
+        if not txn.enabled:
             return self.__wrapped__.send(value)
 
         import aiohttp.web as _web
@@ -123,7 +123,7 @@ class NRTransactionCoroutineWrapper(ObjectProxy):
         txn = self._nr_transaction
 
         # transaction may not be active
-        if not txn._settings:
+        if not txn.enabled:
             return self.__wrapped__.throw(*args, **kwargs)
 
         import aiohttp.web as _web
@@ -139,6 +139,11 @@ class NRTransactionCoroutineWrapper(ObjectProxy):
                 _nr_process_response(response, txn)
             except:
                 pass
+            self._nr_transaction.__exit__(None, None, None)
+            self._nr_request = None
+            raise
+        except asyncio.CancelledError:
+            self._nr_transaction.ignore_transaction = True
             self._nr_transaction.__exit__(None, None, None)
             self._nr_request = None
             raise
@@ -169,7 +174,7 @@ class NRTransactionCoroutineWrapper(ObjectProxy):
         txn = self._nr_transaction
 
         # transaction may not be active
-        if not txn._settings:
+        if not txn.enabled:
             return self.__wrapped__.close()
 
         txn.save_transaction()